requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
TikTokLive>=6.0.0
edge-tts>=6.1.9
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=12.0
pygame>=2.5.0
